Refactor code to separate products concerns with auth concerns
removes products actions, ui, and events and puts them in their own
folder
