+++ assets/scripts/auth/events.js
@@ -43,53 +43,11 @@ const onSignOut = function (event) {
     .catch(ui.signOutFailure)
 }
 
-// end sign-up, sign-in, change-password, sign-out authEvents
-
-const onCreateProduct = (event) => {
-  event.preventDefault()
-  const data = getFormFields(event.target)
-  console.log(data)
-  console.log('onCreateProduct events function reached!')
-  api.createProduct(data)
-    .then(ui.createProductSuccess)
-    .catch(ui.createProductFailure)
-}
-
-const onGetAllProducts = (event) => {
-  event.preventDefault()
-  console.log('onGetAllProducts events function reached!')
-  api.getProduct()
-    .then(ui.getProductSuccess)
-    .catch(ui.getProductFailure)
-}
-
-const onUpdateProduct = (event) => {
-  event.preventDefault()
-  console.log('onUpdateProduct events function reached!')
-  const data = getFormFields(event.target)
-  api.updateProduct(data)
-    .then(ui.onUpdateProductSuccess)
-    .catch(ui.onUpdateProductFailure)
-}
-
-const onDeleteProduct = (event) => {
-  event.preventDefault()
-  console.log('onDeleteProduct events function reached!')
-  const id = getFormFields(event.target)
-  api.deleteProduct(id.product.id)
-    .then(ui.onDeleteProductSuccess)
-    .catch(ui.onDeleteProductFailure)
-}
-
 const addHandlers = function () {
   $('#sign-up').on('submit', onSignUp)
   $('#sign-in').on('submit', onSignIn)
   $('#change-password').on('submit', onChangePassword)
   $('#sign-out').on('submit', onSignOut)
-  $('#create-product').on('submit', onCreateProduct)
-  $('#update-product').on('submit', onUpdateProduct)
-  $('#get-product').on('submit', onGetAllProducts)
-  $('delete-product').on('submit', onDeleteProduct)
 }
 
 module.exports = {
