+++ assets/scripts/auth/ui.js
@@ -77,50 +77,6 @@ const signOutFailure = function (error) {
   $('#message').text('Error on sign out!')
 }
 
-const createProductSuccess = (response) => {
-  console.log('createProductSuccess ui reached!')
-  resetForm($('#create-product'))
-  $('#message').text('You have added a product!')
-}
-
-const createProductFailure = (response) => {
-  $('#message').text('Adding an order failed. Try again.')
-}
-
-const onUpdateProductSuccess = (id) => {
-  console.log('onUpdateProductSuccess ui reached!')
-  resetForm($('#update-book'))
-  $('#message').text('You have updated an order!')
-}
-
-const onUpdateProductFailure = (response) => {
-  console.log('onUpdateProductFailure ui reached!')
-  $('#message').text('Unable to update order! Try again.')
-}
-
-const onDeleteProductSuccess = (id) => {
-  console.log('onDeleteProductSuccess ui reached!')
-  resetForm($('#delete-product'))
-  $('#message').text('You have deleted a product!')
-}
-
-const onDeleteProductFailure = (response) => {
-  console.log('onDeleteProductFailure ui reached!')
-  $('#message').text('Unable to delete a product! Try again.')
-}
-
-const getProductSuccess = (data) => {
-  console.log(data)
-  console.log('ongetProductSuccess ui reached!')
-  resetForm($('#get-product'))
-  $('#message').text('You have get a product!')
-}
-
-const getProductFailure = (response) => {
-  console.log('ongetProductFailure ui reached!')
-  $('#message').text('Unable to get a product! Try again.')
-}
-
 module.exports = {
   signUpSuccess,
   signUpFailure,
@@ -129,13 +85,5 @@ module.exports = {
   changePasswordSuccess,
   changePasswordFailure,
   signOutSuccess,
-  signOutFailure,
-  createProductSuccess,
-  createProductFailure,
-  onUpdateProductSuccess,
-  onUpdateProductFailure,
-  onDeleteProductSuccess,
-  onDeleteProductFailure,
-  getProductSuccess,
-  getProductFailure
+  signOutFailure
 }
