+++ assets/scripts/index.js
@@ -3,6 +3,7 @@
 const setAPIOrigin = require('../../lib/set-api-origin')
 const config = require('./config')
 const authEvents = require('./auth/events')
+const productEvents = require('./products/events')
 
 $(() => {
   setAPIOrigin(location, config)
@@ -16,4 +17,5 @@ $(() => {
 
 $(() => {
   authEvents.addHandlers()
+  productEvents.addHandlers()
 })
